requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiomultiprocess==0.9.0
aiosignal==1.3.1
argcomplete==2.1.1
async-timeout==4.0.2
attrs==22.2.0
bc-detect-secrets==1.4.14
bc-jsonpath-ng==1.5.9
bc-python-hcl2==0.3.51
beautifulsoup4==4.11.2
boto3==1.26.91
botocore==1.29.91
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
ChartAnalysis @ file:///home/vagrant/chart_analysis
click==8.1.3
click-option-group==0.5.5
cloudsplaining==0.5.0
colorama==0.4.6
ConfigArgParse==1.5.3
contextlib2==21.6.0
cyclonedx-python-lib==3.1.5
decorator==5.1.1
deep-merge==0.0.4
docker==6.0.1
dockerfile-parse==2.0.0
dpath==1.5.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
igraph==0.10.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
junit-xml==1.9
lark==1.1.5
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
networkx==2.6.3
packageurl-python==0.10.4
packaging==23.0
pkgutil_resolve_name==1.3.10
ply==3.11
policy-sentry==0.12.6
policyuniverse==1.5.0.20220613
prettytable==3.6.0
pycares==4.3.0
pycep-parser==0.3.9
pycparser==2.21
pyrsistent==0.19.3
pyston==2.3.5
pyston-autoload==2.3.5
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
s3transfer==0.6.0
schema==0.7.5
semantic-version==2.10.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.4
tabulate==0.9.0
termcolor==2.2.0
texttable==1.6.7
toml==0.10.2
tqdm==4.65.0
typing_extensions==4.5.0
unidiff==0.7.5
update-checker==0.18.0
urllib3==1.26.15
wcwidth==0.2.6
websocket-client==1.5.1
yarl==1.8.2
zipp==3.15.0
